Separated forms, todo: add new form components for various circumstances and a dropdown to toggle which forms are shown
